Move SingleMarathonIntegrationTest into its own file.
